Lazy resolve the new text of text edit
- When 'java.completion.lazyResolveTextEdit.enabled' turns on, the completion
  proposal will not resolve the text edit until 'completionItem/resolve'.
